refactor(server): decompose wiring.ts into self-registering service modules
Eliminate wiring.ts as a hot file. All wiring logic now lives in
co-located *.module.ts sidecar files. wiring.ts becomes a 32-line
thin orchestrator that discovers and invokes each module in order.

New files:
- apps/server/src/server/module.ts (ServiceModule interface)
- apps/server/src/services/core.module.ts
- apps/server/src/services/discord.module.ts
- apps/server/src/services/event-subscriptions.module.ts
- apps/server/src/services/infrastructure.module.ts
- apps/server/src/services/lead-engineer.module.ts
- apps/server/src/services/linear-agent.module.ts
- apps/server/src/services/scheduler.module.ts
- apps/server/src/services/worktree-lifecycle.module.ts
- apps/server/src/services/channel-handlers/channel-handlers.module.ts
- apps/server/src/services/escalation-channels/escalation-channels.module.ts

Result: wiring.ts 387→32 lines. New services only need to
create/edit their own *.module.ts — never touch wiring.ts again.
